File state: make symlink deal with other existing types (FIFOs, sockets)
The current checks only take regular files, symlinks and directories
into account. For all other types (FIFOs, sockets) the state doesn't
even attempt to remove them; it simply returns an error that something
exists.

The new code keeps handling of symlinks as-is while generalizing all
other cases by only checking if they exist, and if they do, letting
the 'file.remove' module function handle the removal if 'force' is
true.

The 'file.remove' module function suffered from the same problem. It
was fixed to handle all types in the previous commit.

This is the second part of the fix for #35621.